update6

+++ src/main/java/com/cp/Contests_management/Announcement/AnnouncementRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AnnouncementRepository extends JpaRepository<Announcement,Long> {
+public interface AnnouncementRepository extends JpaRepository<Announcement,Integer> {
 
-    List<Announcement> findByCompetitionId(Long competitionId);
+    List<Announcement> findByCompetitionId(Integer competitionId);
 
 }
