+++ src/main/java/com/cp/Contests_management/Participant/ParticipantRepository.java
@@ -3,8 +3,8 @@ package com.cp.Contests_management.Participant;
 import org.aspectj.apache.bcel.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParticipantRepository extends JpaRepository<Participant, Long> {
+public interface ParticipantRepository extends JpaRepository<Participant, Integer> {
     Participant findByName(String name);
 
-    Participant getParticipantById(Long id);
+    Participant getParticipantById(Integer id);
 }
