+++ src/main/java/com/cp/Contests_management/Problem/ProblemRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProblemRepository extends JpaRepository<Problem, Long> {
+public interface ProblemRepository extends JpaRepository<Problem, Integer> {
     List<Problem> findByName(String name);
 
-    List<Problem> findByCompetitionId(Long competitionId);
+    List<Problem> findByCompetitionId(Integer competitionId);
 }
