+++ src/main/java/com/cp/Contests_management/Submission/SubmissionRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SubmissionRepository extends JpaRepository<Submission, Long> {
-    List<Submission> findByParticipantId(Long participantId);
+public interface SubmissionRepository extends JpaRepository<Submission, Integer> {
+    List<Submission> findByParticipantId(Integer participantId);
 
 
 }
