+++ src/main/java/com/cp/Contests_management/TestCase/TestCaseRepository.java
@@ -4,10 +4,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TestCaseRepository extends JpaRepository<TestCase, Long> {
-    List<TestCase> findByProblemId(Long problemId);
+public interface TestCaseRepository extends JpaRepository<TestCase, Integer> {
+    List<TestCase> findByProblemId(Integer problemId);
 
-    List<TestCase> findByProblemIdAndHiddenTrue(Long problemId);
+    List<TestCase> findByProblemIdAndHiddenTrue(Integer problemId);
 
-    List<TestCase> findByProblemIdAndHiddenFalse(Long problemId);
+    List<TestCase> findByProblemIdAndHiddenFalse(Integer problemId);
 }
